--- src/app/app.server.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ServerModule } from '@angular/platform-server';
-
-import { AppComponent } from './app.component';
-import { AppModule } from './app.module';
-
-@NgModule({
-  imports: [AppModule, ServerModule],
-  bootstrap: [AppComponent]
-})
-export class AppServerModule {}
