--- src/tsconfig.server.json
@@ -1,13 +1,0 @@
-{
-  "extends": "./tsconfig.site.json",
-  "compilerOptions": {
-    "outDir": "./out-tsc/app-server",
-    "module": "commonjs",
-    "types": ["node"],
-    "typeRoots": ["../node_modules/@types"]
-  },
-  "files": ["main.server.ts", "server.ts", "../packages/dev-mode-types.d.ts"],
-  "angularCompilerOptions": {
-    "entryModule": "./app/app.server.module#AppServerModule"
-  }
-}
